refactor: move values to entities
