+++ internal/commands/modify.go
@@ -4,9 +4,9 @@ import (
 	"fmt"
 
 	"github.com/dondakeshimo/todo-cli/internal/entities/task"
+	"github.com/dondakeshimo/todo-cli/internal/entities/reminder"
+	"github.com/dondakeshimo/todo-cli/internal/entities/remindtime"
 	"github.com/dondakeshimo/todo-cli/internal/gateways/json"
-	"github.com/dondakeshimo/todo-cli/internal/values/reminder"
-	"github.com/dondakeshimo/todo-cli/internal/values/remindtime"
 	"github.com/dondakeshimo/todo-cli/pkg/scheduler"
 	"github.com/urfave/cli/v2"
 )
