+++ internal/entities/task/task.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/dondakeshimo/todo-cli/internal/values/reminder"
-	"github.com/dondakeshimo/todo-cli/internal/values/remindtime"
+	"github.com/dondakeshimo/todo-cli/internal/entities/reminder"
+	"github.com/dondakeshimo/todo-cli/internal/entities/remindtime"
 	"github.com/dondakeshimo/todo-cli/pkg/scheduler"
 )
